Add files via upload
This source code parse a post-verilog file to c. 
It also gives the opportunity to the user to simulate the circuit for a random input set or to examine the logical masking of a random Single Event Transients.  When a SET happens the output of a gate changes. The rates show the masking of the exactly circuit  for many SETs in random gates.